@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.11.0 <3.0.0"
+  sdk: ">=2.13.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -44,7 +44,7 @@ dependencies:
   flutter_local_notifications: ^9.1.4
   url_launcher: ^6.0.10
   flutter_launcher_icons: "^0.11.0"
-
+  source_gen: ^1.5.0
 
 
   
