@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 
 class ListModel {
-  String name;
-  Widget widget;
+  String? name;
+  Widget? widget;
 
   ListModel({this.name, this.widget});
 }
