@@ -30,11 +30,11 @@ mixin _$AppStore on AppStoreBase, Store {
   @override
   Color get scaffoldBackground {
     _$scaffoldBackgroundAtom.reportRead();
-    return super.scaffoldBackground;
+    return super.scaffoldBackground!;
   }
 
   @override
-  set scaffoldBackground(Color value) {
+  set scaffoldBackground(Color ?value) {
     _$scaffoldBackgroundAtom.reportWrite(value, super.scaffoldBackground, () {
       super.scaffoldBackground = value;
     });
@@ -45,11 +45,11 @@ mixin _$AppStore on AppStoreBase, Store {
   @override
   Color get backgroundColor {
     _$backgroundColorAtom.reportRead();
-    return super.backgroundColor;
+    return super.backgroundColor!;
   }
 
   @override
-  set backgroundColor(Color value) {
+  set backgroundColor(Color ?value) {
     _$backgroundColorAtom.reportWrite(value, super.backgroundColor, () {
       super.backgroundColor = value;
     });
@@ -61,11 +61,11 @@ mixin _$AppStore on AppStoreBase, Store {
   @override
   Color get backgroundSecondaryColor {
     _$backgroundSecondaryColorAtom.reportRead();
-    return super.backgroundSecondaryColor;
+    return super.backgroundSecondaryColor!;
   }
 
   @override
-  set backgroundSecondaryColor(Color value) {
+  set backgroundSecondaryColor(Color? value) {
     _$backgroundSecondaryColorAtom
         .reportWrite(value, super.backgroundSecondaryColor, () {
       super.backgroundSecondaryColor = value;
@@ -77,11 +77,11 @@ mixin _$AppStore on AppStoreBase, Store {
   @override
   Color get textPrimaryColor {
     _$textPrimaryColorAtom.reportRead();
-    return super.textPrimaryColor;
+    return super.textPrimaryColor!;
   }
 
   @override
-  set textPrimaryColor(Color value) {
+  set textPrimaryColor(Color? value) {
     _$textPrimaryColorAtom.reportWrite(value, super.textPrimaryColor, () {
       super.textPrimaryColor = value;
     });
@@ -93,11 +93,11 @@ mixin _$AppStore on AppStoreBase, Store {
   @override
   Color get appColorPrimaryLightColor {
     _$appColorPrimaryLightColorAtom.reportRead();
-    return super.appColorPrimaryLightColor;
+    return super.appColorPrimaryLightColor!;
   }
 
   @override
-  set appColorPrimaryLightColor(Color value) {
+  set appColorPrimaryLightColor(Color? value) {
     _$appColorPrimaryLightColorAtom
         .reportWrite(value, super.appColorPrimaryLightColor, () {
       super.appColorPrimaryLightColor = value;
@@ -110,11 +110,11 @@ mixin _$AppStore on AppStoreBase, Store {
   @override
   Color get textSecondaryColor {
     _$textSecondaryColorAtom.reportRead();
-    return super.textSecondaryColor;
+    return super.textSecondaryColor!;
   }
 
   @override
-  set textSecondaryColor(Color value) {
+  set textSecondaryColor(Color ?value) {
     _$textSecondaryColorAtom.reportWrite(value, super.textSecondaryColor, () {
       super.textSecondaryColor = value;
     });
@@ -125,11 +125,11 @@ mixin _$AppStore on AppStoreBase, Store {
   @override
   Color get appBarColor {
     _$appBarColorAtom.reportRead();
-    return super.appBarColor;
+    return super.appBarColor!;
   }
 
   @override
-  set appBarColor(Color value) {
+  set appBarColor(Color? value) {
     _$appBarColorAtom.reportWrite(value, super.appBarColor, () {
       super.appBarColor = value;
     });
@@ -140,11 +140,11 @@ mixin _$AppStore on AppStoreBase, Store {
   @override
   Color get iconColor {
     _$iconColorAtom.reportRead();
-    return super.iconColor;
+    return super.iconColor!;
   }
 
   @override
-  set iconColor(Color value) {
+  set iconColor(Color? value) {
     _$iconColorAtom.reportWrite(value, super.iconColor, () {
       super.iconColor = value;
     });
@@ -156,11 +156,11 @@ mixin _$AppStore on AppStoreBase, Store {
   @override
   Color get iconSecondaryColor {
     _$iconSecondaryColorAtom.reportRead();
-    return super.iconSecondaryColor;
+    return super.iconSecondaryColor!;
   }
 
   @override
-  set iconSecondaryColor(Color value) {
+  set iconSecondaryColor(Color? value) {
     _$iconSecondaryColorAtom.reportWrite(value, super.iconSecondaryColor, () {
       super.iconSecondaryColor = value;
     });
@@ -201,7 +201,7 @@ mixin _$AppStore on AppStoreBase, Store {
       AsyncAction('AppStoreBase.toggleDarkMode');
 
   @override
-  Future<void> toggleDarkMode({bool value}) {
+  Future<void> toggleDarkMode({bool? value}) {
     return _$toggleDarkModeAsyncAction
         .run(() => super.toggleDarkMode(value: value));
   }
